chore: refactor street mode selectors
