+++ app/component/itinerary/customizesearch/customize-search.scss
@@ -170,6 +170,7 @@
   color: $black;
   position: relative;
   box-shadow: inset 5px 0 5px 0 $background-color-lighter;
+  font-size: $font-size-normal;
 
   .close-offcanvas {
     position: absolute;
@@ -186,6 +187,12 @@
     color: $black;
   }
 
+  .indent-setting {
+    display: flex;
+    flex-direction: row;
+    margin-left: 4em;
+  }
+
   .toggle-container {
     display: flex;
     flex-direction: row;
@@ -193,6 +200,7 @@
     width: 100%;
     min-height: 3.2em;
     justify-content: space-between;
+    font-size: $font-size-normal;
 
     .toggle-left {
       display: flex;
@@ -269,6 +277,14 @@
     padding-top: 8px;
   }
 
+  .bottom-border {
+    border-bottom: 1px solid $light-gray;
+  }
+
+  .top-border {
+    border-top: 1px solid $light-gray;
+  }
+
   .settings-option-container:last-child {
     border-bottom: 0;
 
@@ -307,60 +323,43 @@
     align-items: center;
   }
 
-  .mode-option-block {
-    min-height: 51.5156px;
-    margin-bottom: 0.1em;
+  .taxi-icon {
+    color: $external-feed-color;
+  }
+
+  .scooter-icon {
+    color: #333;
+  }
+
+  .mode-icon {
+    margin-left: 14.72px;
+    margin-right: 14.72px;
+
+    .subicon-caution .icon {
+      width: 16px;
+      height: 16px;
+      margin-left: 0.1em;
+      margin-top: -0.8em;
+    }
+  }
+
+  label.mode-name {
     display: flex;
-    cursor: pointer;
+    justify-content: space-between;
+    font-weight: $font-weight-medium;
+    font-size: $font-size-small;
+  }
 
-    span {
-      color: #000;
+  .mode-name {
+    width: 100%;
+
+    > span {
+      line-height: 1.1em;
+      margin-top: 1.2em;
     }
 
-    .icon {
-      font-size: 2.5em;
-      margin-top: 0.2em;
-
-      &.taxi-icon {
-        color: $external-feed-color;
-        fill: white;
-      }
-
-      &.scooter-icon {
-        color: #333;
-      }
-    }
-
-    .mode-icon {
-      margin-left: 14.72px;
-      margin-right: 14.72px;
-
-      .subicon-caution .icon {
-        width: 16px;
-        height: 16px;
-        margin-left: 0.1em;
-        margin-top: -0.8em;
-      }
-    }
-
-    label.mode-name {
-      display: flex;
-      justify-content: space-between;
-      font-weight: $font-weight-medium;
-      font-size: $font-size-small;
-    }
-
-    .mode-name {
-      width: 100%;
-
-      > span {
-        line-height: 1.1em;
-        margin-top: 1.2em;
-      }
-
-      .span-bike-not-allowed {
-        font-size: $font-size-xxsmall;
-      }
+    .span-bike-not-allowed {
+      font-size: $font-size-xxsmall;
     }
   }
 }
@@ -405,48 +404,10 @@
   }
 }
 
-.street-mode-selector-panel.customized-settings {
-  flex-direction: column;
-
-  .street-mode-selector-panel-header {
-    margin-bottom: 0.5em;
-    padding-left: 0;
-
-    span {
-      color: rgb(51, 51, 51);
-      text-transform: uppercase;
-      letter-spacing: -0.6pt;
-      font-size: $font-size-normal;
-      font-weight: $font-weight-medium;
-      font-family: $font-family;
-    }
-  }
-
-  .street-mode-selector-panel-buttons button {
-    flex-grow: 1;
-    margin-bottom: 0;
-  }
-
-  .toggle-button-title {
-    margin-bottom: 1.5em;
-  }
-}
-
-.street-modes-container {
-  > div {
-    border-bottom: 1px solid $light-gray;
-  }
-
-  :last-child {
-    border: none;
-  }
-
-  .transport-mode-subheader {
-    border: none;
-  }
-
+.streetmodes {
   .settings-dropdown-wrapper {
-    margin-left: 3.2em;
+    margin-left: 4em;
+    border-bottom: solid 1px $light-gray;
   }
 }
